ARUHA-793 Separate data write layer from stream EventStream logic
