+++ src/main/java/org/zalando/nakadi/service/EventStream.java
@@ -1,12 +1,9 @@
 package org.zalando.nakadi.service;
 
 import com.codahale.metrics.Meter;
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
-import java.nio.charset.StandardCharsets;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.nio.charset.Charset;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -21,33 +18,12 @@ import org.zalando.nakadi.domain.ConsumedEvent;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.repository.EventConsumer;
 import org.zalando.nakadi.util.FeatureToggleService;
-import org.zalando.nakadi.view.Cursor;
 import static java.lang.System.currentTimeMillis;
 import static java.util.function.Function.identity;
 
 public class EventStream {
 
     private static final Logger LOG = LoggerFactory.getLogger(EventStream.class);
-    public static final String BATCH_SEPARATOR = "\n";
-    public static final Charset UTF8 = Charset.forName("UTF-8");
-
-    private static final byte B_BATCH_SEPARATOR = '\n';
-    private static final byte[] B_CURSOR_PARTITION_BEGIN =
-        "{\"cursor\":{\"partition\":\"".getBytes(StandardCharsets.UTF_8);
-    private static final byte[] B_OFFSET_BEGIN = "\",\"offset\":\"".getBytes(StandardCharsets.UTF_8);
-    private static final byte[] B_CURSOR_PARTITION_END = "\"}".getBytes(StandardCharsets.UTF_8);
-    private static final byte[] B_CLOSE_CURLY_BRACKET = "}".getBytes(StandardCharsets.UTF_8);
-    private static final byte[] B_EVENTS_ARRAY_BEGIN = ",\"events\":[".getBytes(StandardCharsets.UTF_8);
-    private static final byte B_COMMA_DELIM = ',';
-    private static final byte B_CLOSE_BRACKET = ']';
-    private static final int B_FIXED_BYTE_COUNT = B_CURSOR_PARTITION_BEGIN.length
-        + B_OFFSET_BEGIN.length
-        + B_CURSOR_PARTITION_END.length
-        + B_EVENTS_ARRAY_BEGIN.length
-        + 1 // B_COMMA_DELIM
-        + 1 // B_CLOSE_BRACKET
-        + B_CLOSE_CURLY_BRACKET.length
-        + 1; //B_BATCH_SEPARATOR
 
     private final OutputStream outputStream;
     private final EventConsumer eventConsumer;
@@ -171,71 +147,15 @@ public class EventStream {
                 .collect(Collectors.toMap(identity(), valueFunction));
     }
 
-    public static String createStreamEvent(final Cursor cursor, final List<String> events) {
-        final StringBuilder builder = new StringBuilder()
-                .append("{\"cursor\":{\"partition\":\"").append(cursor.getPartition())
-                .append("\",\"offset\":\"").append(cursor.getOffset()).append("\"}");
-        if (!events.isEmpty()) {
-            builder.append(",\"events\":[");
-            events.forEach(event -> builder.append(event).append(","));
-            builder.deleteCharAt(builder.length() - 1).append("]");
-        }
-
-        builder.append("}").append(BATCH_SEPARATOR);
-
-        return builder.toString();
-    }
-
     private void sendBatch(final NakadiCursor topicPosition, final List<String> currentBatch)
             throws IOException {
-        if(featureToggleService.isFeatureEnabled(FeatureToggleService.Feature.SEND_BATCH_VIA_OUTPUT_STREAM)) {
-            writeStreamEvent(outputStream, cursorConverter.convert(topicPosition), currentBatch);
-        } else {
-            // create stream event batch for current partition and send it; if there were
-            // no events, it will be just a keep-alive
-            final String streamEvent = createStreamEvent(cursorConverter.convert(topicPosition), currentBatch);
-            final byte[] batchBytes = streamEvent.getBytes(UTF8);
-            outputStream.write(batchBytes);
-            bytesFlushedMeter.mark(batchBytes.length);
-            outputStream.flush();
-        }
+        final int bytesWritten = EventStreamWriter.getWriter(
+                featureToggleService.isFeatureEnabled(FeatureToggleService.Feature.SEND_BATCH_VIA_OUTPUT_STREAM))
+                .writeBatch(outputStream, cursorConverter.convert(topicPosition), currentBatch);
+        bytesFlushedMeter.mark(bytesWritten);
+        outputStream.flush();
     }
 
-    @VisibleForTesting
-    void writeStreamEvent(final OutputStream os, final Cursor cursor, final List<String> events)
-        throws IOException {
-
-        int byteCount = B_FIXED_BYTE_COUNT;
-
-        os.write(B_CURSOR_PARTITION_BEGIN);
-        final byte[] partition = cursor.getPartition().getBytes(StandardCharsets.UTF_8);
-        os.write(partition);
-        byteCount += partition.length;
-        os.write(B_OFFSET_BEGIN);
-        final byte[] offset = cursor.getOffset().getBytes(StandardCharsets.UTF_8);
-        os.write(offset);
-        byteCount += offset.length;
-        os.write(B_CURSOR_PARTITION_END);
-        if (!events.isEmpty()) {
-            os.write(B_EVENTS_ARRAY_BEGIN);
-            for (int i = 0; i < events.size(); i++) {
-                final byte[] event = events.get(i).getBytes(StandardCharsets.UTF_8);
-                os.write(event);
-                byteCount += event.length;
-                if(i < (events.size() - 1)) {
-                    os.write(B_COMMA_DELIM);
-                }
-                else {
-                    os.write(B_CLOSE_BRACKET);
-                }
-            }
-        }
-        os.write(B_CLOSE_CURLY_BRACKET);
-        os.write(B_BATCH_SEPARATOR);
-
-        bytesFlushedMeter.mark(byteCount);
-        os.flush();
-    }
 
     public void close() throws IOException {
         this.eventConsumer.close();
