+++ src/test/java/org/zalando/nakadi/service/EventStreamTest.java
@@ -6,26 +6,37 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
-import java.util.Map;
-import org.zalando.nakadi.config.JsonConfig;
-import org.zalando.nakadi.util.FeatureToggleService;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import static java.util.Collections.nCopies;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import static java.util.Optional.empty;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import static junit.framework.TestCase.assertSame;
+import static junit.framework.TestCase.fail;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.arrayWithSize;
+import static org.hamcrest.core.Is.is;
 import org.junit.Assert;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+import org.zalando.nakadi.config.JsonConfig;
 import org.zalando.nakadi.domain.ConsumedEvent;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.Timeline;
@@ -33,28 +44,14 @@ import org.zalando.nakadi.exceptions.NakadiException;
 import org.zalando.nakadi.repository.db.EventTypeCache;
 import org.zalando.nakadi.repository.kafka.KafkaCursor;
 import org.zalando.nakadi.repository.kafka.NakadiKafkaConsumer;
+import static org.zalando.nakadi.service.EventStreamWriter.BATCH_SEPARATOR;
 import org.zalando.nakadi.service.converter.CursorConverterImpl;
 import org.zalando.nakadi.service.timeline.TimelineService;
-import org.zalando.nakadi.view.Cursor;
-import static java.util.Collections.nCopies;
-import static java.util.Optional.empty;
-import static junit.framework.TestCase.assertSame;
-import static junit.framework.TestCase.fail;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.arrayWithSize;
-import static org.hamcrest.core.Is.is;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
-import static org.mockito.Matchers.anyInt;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.zalando.nakadi.service.EventStream.BATCH_SEPARATOR;
+import org.zalando.nakadi.util.FeatureToggleService;
 import static org.zalando.nakadi.utils.TestUtils.createFakeTimeline;
 import static org.zalando.nakadi.utils.TestUtils.randomString;
 import static org.zalando.nakadi.utils.TestUtils.waitFor;
+import org.zalando.nakadi.view.Cursor;
 import static uk.co.datumedge.hamcrest.json.SameJSONAs.sameJSONAs;
 
 public class EventStreamTest {
@@ -379,17 +376,6 @@ public class EventStreamTest {
     @Test
     public void testWriteStreamEvent() {
 
-        final Meter meter = mock(Meter.class);
-
-        final EventStream eventStream =
-            new EventStream(null,
-                null,
-                null,
-                null,
-                null,
-                meter,
-                null);
-
         final ByteArrayOutputStream baos = new ByteArrayOutputStream();
         final Cursor cursor = new Cursor("22", "000000000000000023");
         final ArrayList<String> events = Lists.newArrayList(
@@ -398,7 +384,7 @@ public class EventStreamTest {
             "{\"e\":\"f\"}");
 
         try {
-            eventStream.writeStreamEvent(baos, cursor, events);
+            EventStreamWriterString.INSTANCE.writeBatch(baos, cursor, events);
             final Map<String, Object> batch =
                 mapper.readValue(baos.toString(), new TypeReference<Map<String, Object>>() {});
 
@@ -414,8 +400,6 @@ public class EventStreamTest {
             assertEquals("d", eventsM.get(1).get("c"));
             assertEquals("f", eventsM.get(2).get("e"));
 
-            verify(meter, times(1)).mark(anyInt());
-
         } catch (IOException e) {
             fail(e.getMessage());
         }
@@ -425,22 +409,12 @@ public class EventStreamTest {
     @Test
     public void testWriteStreamEventEmptyBatchProducesNoEventArray() {
 
-        final Meter meter = mock(Meter.class);
-
-        final EventStream eventStream =
-            new EventStream(null,
-                null,
-                null,
-                null,
-                null,
-                meter,
-                null);
         final ByteArrayOutputStream baos = new ByteArrayOutputStream();
         final Cursor cursor = new Cursor("11", "000000000000000012");
         final ArrayList<String> events = Lists.newArrayList();
 
         try {
-            eventStream.writeStreamEvent(baos, cursor, events);
+            EventStreamWriterString.INSTANCE.writeBatch(baos, cursor, events);
             final String json = baos.toString();
 
             assertEquals("{\"cursor\":{\"partition\":\"11\",\"offset\":\"000000000000000012\"}}\n", json);
@@ -455,8 +429,6 @@ public class EventStreamTest {
             final List<Map<String, String>> eventsM = (List<Map<String, String>>) batch.get("events");
             // expecting events not to be written as an empty array
             assertNull(eventsM);
-
-            verify(meter, times(1)).mark(anyInt());
 
         } catch (IOException e) {
             fail(e.getMessage());
